moving methods around to clean things up.
